Reconstructed process detection to work with a new IProcessService interface
